Code improvements, reduced icon sizes, bug fixes

+++ src/markdownparser.cpp
@@ -54,7 +54,7 @@ auto Parser::toHtml(const QString &in, const int dia) -> QString
 
     const QByteArray array = in.toUtf8(); // Use UTF-8 for better support
     QByteArray out = templateArray;
-    out.reserve(int(array.size() *1.28) + 115);
+    out.reserve(array.size() *1.28 + 115);
 
     md_html(array.constData(), array.size(), &captureHtmlFragment, &out,
             parser_flags, MD_HTML_FLAG_SKIP_UTF8_BOM);
